% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{art_mean}
\alias{art_mean}
\title{Function Test}
\usage{
art_mean(x)
}
\arguments{
\item{x}{is a numeric vector}
}
\value{
A numeric result
}
\description{
A test function to calculate manual arithmetic mean.
}
\examples{
x <- c(1,2,3)
art_mean(x)
}
